requirements.txt

pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.2
umap-learn>=0.5.3
hdbscan>=0.8.33
openai>=1.0.0
scipy>=1.10.0
joblib>=1.3.0
PyYAML>=6.0
matplotlib
seaborn